Set up error reporting capability
Previously, it was assumed that `run` would execute the provided source
code without errors, but this will not always be the case. For example,
we might encounter syntax errors during scanning.

This patch adds the capacity to encounter and report errors - both when
running a Lox source file, and in the REPL.

When an error is encountered in the REPL, it does not terminate the program.

Only one kind of error (SyntaxError) is added as part of this patch.